requirements.txt
PyYAML>=6.0
filelock>=3.16.0
tzdata>=2024.1
python-dotenv>=1.0.1
alpaca-py>=0.30.0
pytz>=2024.1
pydantic>=2.0
supabase<2
pytest>=8.3.0
streamlit>=1.28.0
requests>=2.31.0
